@@ -4,18 +4,18 @@ buildscript {
 
     ext.versions = [
             'minSdk'        : 15,
-            'compileSdk'    : 25,
-            'targetSdk'     : 25,
+            'compileSdk'    : 28,
+            'targetSdk'     : 28,
 
-            'androidTools'  : '27.0.3',
+            'androidTools'  : '29.0.2',
             'supportLibrary': '25.3.1',
     ]
 
     ext.deps = [
             android : [
                     support: [
-                            'appcompat' : "com.android.support:appcompat-v7:${versions.supportLibrary}",
-                            'constraint': 'com.android.support.constraint:constraint-layout:1.0.2',
+                            'appcompat' : 'androidx.appcompat:appcompat:1.0.2',
+                            'constraint': 'androidx.constraintlayout:constraintlayout:1.1.3',,
                             test        : [
                                     'espresso': 'com.android.support.test.espresso:espresso-core:2.2.2',
                                     'runner'  : 'android.support.test.runner.AndroidJUnitRunner'
@@ -28,9 +28,10 @@ buildscript {
 
     repositories {
         jcenter()
+        google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.3.2'
+        classpath 'com.android.tools.build:gradle:3.5.0'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -40,6 +41,7 @@ buildscript {
 allprojects {
     repositories {
         jcenter()
+        google()
     }
 }
 
