@@ -7,7 +7,8 @@ import android.content.ServiceConnection;
 import android.os.Bundle;
 import android.os.IBinder;
 import android.os.RemoteException;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.shadowwingz.androidlifedemo.R;
 import com.shadowwingz.androidlifedemo.utils.LogUtil;
