@@ -1,7 +1,8 @@
 package com.shadowwingz.androidlifedemo.customviewdemo.conpon;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.shadowwingz.androidlifedemo.R;
 
