@@ -3,9 +3,10 @@ package com.shadowwingz.androidlifedemo;
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.SystemClock;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.shadowwingz.androidlifedemo.binderdemo.BookManagerActivity;
 import com.shadowwingz.androidlifedemo.customviewdemo.CustomViewActivity;
