@@ -1,10 +1,11 @@
 package com.shadowwingz.androidlifedemo.eventdemo.view;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
 import android.view.View;
+
+import androidx.annotation.Nullable;
 
 import com.shadowwingz.androidlifedemo.utils.LogUtil;
 
