@@ -1,7 +1,8 @@
 package com.shadowwingz.androidlifedemo.setcontentviewdemo;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.shadowwingz.androidlifedemo.R;
 
